+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+- Exposed AdQuartile event
+
+### Fixed
+- Player pausing indefinitely when TruexConfiguration is null and a TrueX ad is found
+
 ## [1.5.0]
 
 ### Changed